\name{db.data.frame-class}
\Rdversion{1.1}
\docType{class}
\alias{db.data.frame-class}

\title{Class \code{"db.data.frame"}}
\description{
A concise (1-5 lines) description of what the class is. ~~
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "db.data.frame" in the signature.
}
\references{
put references to the literature/web site here~~
}
\author{
who you are~~
}
\note{
further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  objects to See Also
  %as \code{\link{~~fun~~}}, ~~~
%or \code{\linkS4class{CLASSNAME}} for links to other classes ~~~
}
\examples{
showClass("db.data.frame")
}
\keyword{classes}
